rename: 타임라인 대소문자 변경
